Fix for "rawRID"s that do not need to be truncated
Adds a conditional that prevents errors when esenutl.py returns an RID instead of the expected SID.

esenutl.py usually returns ATTr589970 (the "rawRID") as a full-length SID, but sometimes an 8 character string which is just the RID. This resulted in errors for some Windows domains because it truncated to blank strings, which couldn't convert to integers. This patch checks whether the truncation is needed.